hello-spring-framework-mvc/src/test/java/com/sme/spring/mvc/controller/HomeControllerTest.java
package com.sme.spring.mvc.controller;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.result.MockMvcResultMatchers;

/**
 * Unit tests of {@link HomeController}.
 */
@SpringBootTest
@AutoConfigureMockMvc
public class HomeControllerTest extends Assertions
{
    @Autowired
    private MockMvc mockMvc;

    @Test
    void testHome() throws Exception
    {
        mockMvc.perform(MockMvcRequestBuilders.get("/"))
                .andExpect(MockMvcResultMatchers.status().isOk())
                .andExpect(MockMvcResultMatchers.model().attributeExists("articles"))
                .andExpect(MockMvcResultMatchers.view().name("home"));
    }
}
